feat: like, unlike model and schema

+++ routers/likedRouter.js
@@ -1,30 +1,33 @@
 const express = require("express")
 const router = express.Router()
-const { Unliked } = require('../models/unlikedModel.js')
 const { Liked } = require('../models/likedModel.js')
+const { Unliked } = require('../models/unlikedModel.js')
 
 router.route('/')
 .get( async (req, res) => {
 
-  const unliked = await Unliked.find() 
-  res.status(200).send(unliked)
+  const liked = await Liked.find() 
+  res.status(200).send(liked)
+  
+
 })
 .post( async (req, res) => {
 
    try {
     const {name, date, thumbnail, url, duration, description, category, videoId, subCategory} = req.body
 
-    const alreadyPresentinUnliked = await Unliked.find({ videoId: videoId})
-    const alreadyPresentinLiked = await Liked.find({ videoId: videoId})
-    if(alreadyPresentinUnliked)
-    {
-      await Unliked.deleteOne({videoId: videoId})
-   }
-       if(alreadyPresentinLiked)
+    const alreadyPresentInLiked = await Liked.find({ videoId: videoId})
+    const alreadyPresentInUnliked = await Unliked.find({videoId: videoId})
+
+    if(alreadyPresentInLiked)
     {
       await Liked.deleteOne({videoId: videoId})
-   }
-    const unliked = new Unliked( {
+    }
+    if(alreadyPresentInUnliked)
+    {
+      await Unliked.deleteOne({videoId: videoId})
+    }
+    const liked = new Liked( {
       name: name, 
       date: date, 
       thumbnail: thumbnail, 
@@ -36,9 +39,9 @@ router.route('/')
       subCategory: subCategory
 
     })
-    const savedUnliked = await unliked.save()
-    console.log(savedUnliked)
-    res.json({success: true, data: alreadyPresent})
+    const savedLiked = await liked.save()
+    console.log(savedLiked)
+    res.json({success: true, data: alreadyPresentInLiked})
        } 
    catch(err){
         res.json({success: false, errorrrrr: err})
